+++ packages/eyes-sdk-core/lib/config/IosVersion.js
@@ -1,13 +1,14 @@
 const Enum = require('../utils/Enum')
 
 /**
- * @typedef {string} IosVersion
+ * @typedef {import('../utils/Enum').EnumValues<typeof IosVersions>} IosVersion
  */
 
 /**
  * iOS version for visual-grid rendering
  */
 const IosVersions = Enum('IosVersion', {
+  /** @type {'latest'} */
   Latest: 'latest',
 })