+++ packages/eyes-webdriverio-5/src/SpecWrappedElement.js
@@ -2,9 +2,14 @@ const {TypeUtils} = require('@applitools/eyes-sdk-core')
 const LegacySelector = require('./LegacySelector')
 
 /**
+ * @typedef {{ELEMENT?: string, 'element-6066-11e4-a52e-4f735466cecf'?: string}} WDIOPlainElement
+ * @typedef {{elementId: string, selector: string} & WDIOPlainElement} WDIOElement
+ */
+
+/**
  * @typedef {import('./SpecWrappedDriver').Driver} Driver
  * @typedef {string|Function|import('./LegacySelector')} Selector
- * @typedef {import('webdriverio').Element|{ELEMENT: string, 'element-6066-11e4-a52e-4f735466cecf': string}} Element
+ * @typedef {WDIOElement|WDIOPlainElement} Element
  *
  * @typedef {import('@applitools/eyes-sdk-core').SpecElement<Driver, Element, Selector>} WDIOSpecElement
  */
@@ -63,14 +68,11 @@ function isStaleElementReferenceResult(result) {
   return errOrResult instanceof Error && errOrResult.name === 'stale element reference'
 }
 
-/** @type {WDIOSpecElement} */
-module.exports = {
-  isCompatible,
-  isSelector,
-  toSupportedSelector,
-  toEyesSelector,
-  extractId,
-  extractElement,
-  extractSelector,
-  isStaleElementReferenceResult,
-}
+exports.isCompatible = isCompatible
+exports.isSelector = isSelector
+exports.toSupportedSelector = toSupportedSelector
+exports.toEyesSelector = toEyesSelector
+exports.extractId = extractId
+exports.extractElement = extractElement
+exports.extractSelector = extractSelector
+exports.isStaleElementReferenceResult = isStaleElementReferenceResult