+++ packages/eyes-webdriverio-5/src/WDIOWrappedElement.js
@@ -1,6 +1,6 @@
 const {EyesWrappedElement} = require('@applitools/eyes-sdk-core')
 const LegacyWrappedElement = require('./LegacyWrappedElement')
-
+/** @type {import('./SpecWrappedElement'.WDIOSpecElement)} */
 const SpecWrappedElement = require('./SpecWrappedElement')
 
 const WDIOWrappedElement = EyesWrappedElement.specialize(SpecWrappedElement)