+++ packages/sdk-coverage-tests/js/supported-tests.js
@@ -311,6 +311,7 @@ module.exports = [
   {name: 'TestWindowWithModal_Fully', executionMode: {isScrollStitching: true}},
   {name: 'TestTooBigViewportSize', executionMode: {isCssStitching: true}},
   {name: 'TestSetViewportSize', executionMode: {isCssStitching: true}},
+  {name: 'TestSetViewportSizeEdge', executionMode: {isCssStitching: true}},
   {name: 'TestVisualLocators', executionMode: {isCssStitching: true}},
   {name: 'TestVisualLocators', executionMode: {isVisualGrid: true}},
   {name: 'TestMobileNativeRegion', executionMode: {isCssStitching: true}},